requirements.txt
google-generativeai>=0.7.2

streamlit>=1.28.0

pandas>=1.5.0
numpy>=1.21.0

matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0


pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

jupyter
ipykernel
